the first proto loop in the daemon for processing bitshares asset deposits, getting the sender's pub key and turning that into a bitcoin address

+++ BitsharesRpc/Responses.cs
@@ -25,7 +25,7 @@ namespace BitsharesRpc
 
 	public class GetInfoResponse
 	{
-		public ulong blockchain_head_block_num;
+		public uint blockchain_head_block_num;
 		public string blockchain_head_block_age;
 		public DateTime? blockchain_head_block_timestamp;
 		public string blockchain_average_delegate_participation;
